many fixes: textField check for city name, refactoring

+++ WeatherOnTheMap/Models/AllCurrentWeather.swift
@@ -75,7 +75,7 @@ struct Main: Decodable {
 }
 
 struct Wind: Decodable {
-	let deg: Int?
+	let deg: Double?
 	let speed: Double?
 }
